refactor(execute/fill): move common fixtures and hooks in their own plugin
This moves shared fixtures and hooks to their own plugin, that both commands can load as a plugin. This is a simple move; no deeper refactor required
