+++ pytest-execute-hive.ini
@@ -13,6 +13,7 @@ addopts =
     -p pytest_plugins.solc.solc
     -p pytest_plugins.execute.rpc.hive
     -p pytest_plugins.execute.execute
+    -p pytest_plugins.shared.execute_fill
     -p pytest_plugins.forks.forks
     -p pytest_plugins.spec_version_checker.spec_version_checker
     -p pytest_plugins.pytest_hive.pytest_hive
