+++ pytest-execute.ini
@@ -12,6 +12,7 @@ addopts =
     -p pytest_plugins.execute.pre_alloc
     -p pytest_plugins.solc.solc
     -p pytest_plugins.execute.execute
+    -p pytest_plugins.shared.execute_fill
     -p pytest_plugins.execute.rpc.remote_seed_sender
     -p pytest_plugins.execute.rpc.remote
     -p pytest_plugins.forks.forks
